Base histories

+++ api/src/token/token.service.ts
@@ -26,7 +26,13 @@ export class TokenService {
   }
 
   async removeToken(refreshToken: string) {
-    await this.tokenRepository.destroy({ where: { token: refreshToken } });
+    if (refreshToken) {
+      await this.tokenRepository.destroy({ where: { token: refreshToken } });
+    }
+  }
+
+  async removeTokenByUserId(id: number) {
+    await this.tokenRepository.destroy({ where: { userId: id } });
   }
 
   async refreshToken(refreshToken: string) {}
